fire 'loadstart' event when youtube videos load (#21)
This is a step toward making youtube embeds better mimick HTML5 <video>
elements. The specific issue prompting this was that fullscreen wasn't
working for YouTube vids, and it was because the fullscreen mode
detection never happened, because 'loadstart' never fired.

This also fixes `test/test-youtube.html`, which wasn't rendering the
events table because it was expecting a `video` (or `audio`) tag, when
in fact YouTube renders an `iframe`.