Size time slider only based on visible controls
The calculation for sizing the time slider includes a check for whether
the last control in the list wrapped to the next line if the slider was
too long, and to reduce the slider size incrementally till that was no
longer the case. If the last control is set to `display: none`, this
calculation becomes unpredictable and in some cases continues till the
slider is 1px wide.

Hidden controls shouldn’t be affecting the slider size.