fix keyboard accessibility for speed selector
in commit 15bb49394adde366cdd3ef2da72395454a45d18a, `display: none` was
removed from the css for `.mejs-controls .mejs-speed-button
.mejs-speed-selector` in the build file, but not the source file, so the
next time the css got build, the change was reverted.

my understanding of the current best practice for player button menu
accessibility is that neither display or visibility is used to hide the
menu directly, but that it is created with the `mejs-offscreen` and then
we toggle that class on or off to display/hide the menu. this change
puts this menu in line with other menus that use this pattern (tracks,
sourcechooser)