sanitize html outside of active DOM, fixes ARC-2690
mounting the given HTML in an element in the active DOM was actually
causing the browser to execute that new tree. this allowed for
injection attacks such as:

<img src=x onerror=console.info(0)>

even though that element would have been removed after mounting.

instead let’s just parse and sanitize the html outside of the active
DOM.