Remove dfxp code now that we have disabled it
In https://github.com/instructure/mediaelement/commit/5f3c0d6fd00a653db
We removed the ability to show dfxp caption tracks (to fix an xss bug).
But we still left the code to actually do the dfxp parsing behind.
That is just dead code now and is needless bytes we send to the browser

This just removes it