Wait for seek to complete (refs ARC-4646) (#49)
The YouTube API doesn't provide any indication that seeking has
completed when calling the seekTo method. MediaElement currently gets
around this by waiting 100ms before firing events, but some videos can
take much longer, especially if the network connection is poor. This
results in the UI not updating to the expected position when clicking,
seeming to be an event behind in our observation.

This change replaces the static timeout with a polling approach that
improves the situation when the video is paused and a slow seek occurs.